requirements.txt
streamlit==1.32.2
sentence-transformers==3.1.1
transformers==4.46.0
faiss-cpu==1.7.4
requests
